EmbeddedGraphs.jl

EmbeddedGraphs.jl

Documentation for EmbeddedGraphs.jl

A simple package that provides support for embedded graphs.

EmbeddedGraphs.EmbeddedGraph — Type.

EmbeddedGraph{T<:Integer} <: AbstractEmbeddedGraph{T}

Embedded Graph

source

EmbeddedGraphs.EmbeddedGraph — Method.

Constructor functions for the EmbeddedGraph structure

source

EmbeddedGraphs.EuclideanGraph — Type.

EuclideanGraph{T<:Integer} <: AbstractEmbeddedGraph{T}

Euclidean Graph

source

EmbeddedGraphs.EuclideanGraph — Method.

Constructor functions for the EuclideanGraph structure

source

Base.getindex — Method.

Extends Base.getindex function to be able to call EG[i,j]. Returns the distance of the two vertices in the metric of the graph.

source

Base.getindex — Method.

Extends Base.getindex function to be able to call EG[i,j]. Returns the distance of the two vertices in the metric of the graph.

source

Base.rand — Method.

Extends the Base.rand function to work with a SimpleEdgeIter

source

EmbeddedGraphs.characteristic_length — Method.

Calculate the characteristic length (average shortest path) of a graph.

source

EmbeddedGraphs.characteristic_length_ER — Method.

Average path length of ER network as in 10.1103/PhysRevE.70.056110 .

source

EmbeddedGraphs.degree_array — Method.

Returns the degree histogram in an array.

source

EmbeddedGraphs.global_clustering_coefficient_ER — Method.

clustering coefficient of an erdos renyi graph is equal to <k>/(n-1), because the probability of any vertex to be connected to any other vertex is the same and equal to the existence of any link.

source

EmbeddedGraphs.global_clustering_coefficient_Latt — Method.

clustering coefficient of a ring lattice with average degree k = 2m/n. Only graphs with even k can be created, therefore the upper and lower even k lattices are average with weights

source

EmbeddedGraphs.largest_component! — Method.

Changes the given graph, so that only the largest component remains

source

EmbeddedGraphs.local_clustering_histogram — Method.

Returns the local clustering coefficient as a histogram in an array.

source

EmbeddedGraphs.random_geometric_graph! — Method.

random_geometric_graph!(eg, radius; dist_func=Nothing)

From an existing embedded complete graph eg, create a random geometric graph by removing edges between vertices that are more than radius apart. This in-place graph generator offers more flexibility with respect to the properties of the resulting EmbeddedGraph instance eg. For instance, the embedding metric determining the values of weights(eg) can be distinct from the distance function dist_func that is used to create the random geometric graph.

Ref.: Penrose, Mathew. Random geometric graphs. Vol. 5. Oxford university press, 2003.

Optional Arguments

dist_func: distance function used to construct the random geometric graph. It can differ from the distance attribute of eg to allow for independent edge weights. Defaults to eg.distance when dist_func=Nothing is passed.

Examples

julia> egh = EmbeddedGraph(complete_graph(n), pos, hamming);

julia> random_geometric_graph!(egh, rad; dist_func=euclidean)

julia> egh.graph
{50, 71} undirected simple Int64 graph

source

EmbeddedGraphs.random_geometric_graph — Method.

random_geometric_graph(n, radius; dim=2, pos=Nothing, embedding_metric=Nothing, dist_func=Nothing)

Create a random geometric graph with n vertices that are connected when their distance is less than radius. The vertex locations are randomly drawn from a uniform distribution on the unit interval in each dimension.

Ref.: Penrose, Mathew. Random geometric graphs. Vol. 5. Oxford university press, 2003.

Optional Arguments

dim: embedding dimension. Defaults to dim=2.

pos: vertex positions can be given as a list of n points with dim coordinates. Defaults to pos=Nothing.

dist_func: distance function used to construct the random geometric graph. It can differ from the distance attribute of eg to allow for independent edge weights. Defaults to eg.distance when dist_func=Nothing is passed.

Examples

julia> random_geometric_graph(50, 0.2).graph
{50, 105} undirected simple Int64 graph

source

EmbeddedGraphs.rem_vertices! — Method.

Removes multiple vertices with given Indices at once

source

EmbeddedGraphs.small_world_ness_HG — Method.

Implementation of the Small-World-Ness measure by M. D. Humphries and K. Gurney. doi:10.1371/journal.pone.0002051

S = γC / γL, where γC = GCC(graph) / GCC(random graph) , and γL = CL(graph) / CL(random graph)

GCC := Global Clustering Coefficient CL := arachteristic Length

source

EmbeddedGraphs.small_world_ness_Telesford — Method.

Small-world-ness measure as in 10.1089/brain.2011.0038

source

EmbeddedGraphs.vertices_loc — Method.

Functions return an array with the value of the vertices in the given spatial direction 1 == x direction, 2 == y direction, 3 == z direction ...

source

EmbeddedGraphs.wiring_length — Method.

Calculate the sum of all the weights of the edges of a graph.

source

LightGraphs.SimpleGraphs.add_vertex! — Method.

Adds vertex in the given graph. Position in the form [x,y,z,...] needed.

source

LightGraphs.SimpleGraphs.rem_vertex! — Method.

Extends LightGraphs function and deletes also the element in the
EG.vertexpos array. rem_vertex! in LightGraphs swaps the vertex v with
the last vertex and then uses pop! on the list to delete the last element.

source

LightGraphs.add_vertices! — Method.

Adds multiple vertices in the given graph at once. Position Array is needed.

source

LightGraphs.edges — Method.

Extends basic LightGraphs functions to work with AbstractEmbeddedGraphs

source

LightGraphs.weights — Method.

Weights function gives back the weightmatrix of the graph.
It can be choosen whether the matrix should be dense, i.e. the distance
between every vertex is inserted, or sparse, i.e. only the distance
between connected vertices is given back.

source

Index

EmbeddedGraphs.EmbeddedGraph

EmbeddedGraphs.EmbeddedGraph

EmbeddedGraphs.EuclideanGraph

EmbeddedGraphs.EuclideanGraph

Base.getindex

Base.getindex

Base.rand

EmbeddedGraphs.characteristic_length

EmbeddedGraphs.characteristic_length_ER

EmbeddedGraphs.degree_array

EmbeddedGraphs.global_clustering_coefficient_ER

EmbeddedGraphs.global_clustering_coefficient_Latt

EmbeddedGraphs.largest_component!

EmbeddedGraphs.local_clustering_histogram

EmbeddedGraphs.random_geometric_graph

EmbeddedGraphs.random_geometric_graph!

EmbeddedGraphs.rem_vertices!

EmbeddedGraphs.small_world_ness_HG

EmbeddedGraphs.small_world_ness_Telesford

EmbeddedGraphs.vertices_loc

EmbeddedGraphs.wiring_length

LightGraphs.SimpleGraphs.add_vertex!

LightGraphs.SimpleGraphs.rem_vertex!

LightGraphs.add_vertices!

LightGraphs.edges

LightGraphs.weights